sys-kernel/gentoo-sources: add if use build-kernel to pkg_postrm

+++ sys-kernel/gentoo-sources/gentoo-sources-5.10.77-r1.ebuild
@@ -766,7 +766,7 @@ pkg_postrm() {
 
 	# clean-up the generated initramfs for this kernel ...
 	# ... firstly checking that it exists, otherwise we're done.
-	if [[ -f "${EROOT}"/boot/initramfs-${KERNEL_FULL_VERSION} ]]; then
+	if use build-kernel && [[ -f "${EROOT}"/boot/initramfs-${KERNEL_FULL_VERSION} ]]; then
 		rm -f "${EROOT}"/boot/initramfs-${KERNEL_FULL_VERSION} || die "failed to remove initramfs-${KERNEL_FULL_VERSION}"
 	fi
 }
